Qala/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Deprecated==1.2.18
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.25.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
packaging==25.0
phonenumbers==8.13.55
pkce==1.0.3
propcache==0.4.1
pycparser==3.0
pycryptodome==3.20.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
requests==2.32.5
requests-file==3.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
supertokens_python==0.31.0
tldextract==5.3.1
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==1.17.3
yarl==1.23.0
